+++ patterns/header-light-with-hamburger-menu.php
@@ -19,7 +19,7 @@
 <!-- wp:site-title {"level":0,"className":"has-bottom-border","fontSize":"medium"} /-->
 
 <!-- wp:social-links {"iconColor":"main","iconColorValue":"#000000","showLabels":false,"className":"is-style-logos-only","style":{"spacing":{"blockGap":{"top":"var:preset|spacing|small","left":"var:preset|spacing|small"}}}} -->
-<ul class="wp-block-social-links has-icon-color is-style-logos-only"><!-- wp:social-link {"url":"#","service":"facebook"} /-->
+<ul class="wp-block-social-links has-icon-color is-style-logos-only"><!-- wp:social-link {"url":"#","service":"mastodon"} /-->
 <!-- wp:social-link {"url":"#","service":"instagram"} /-->
 <!-- wp:social-link {"url":"#","service":"bluesky"} /--></ul>
 <!-- /wp:social-links --></div>